Nebula Version 74

+++ updater.go
@@ -218,12 +218,13 @@ func main() {
 
 				fmt.Printf("🟣The previous version is %s🟣\n", previousVersion)
 
-				previousVersionPath := filepath.Join(SALTOLocation, previousVersion)
-
-				err = os.RemoveAll(previousVersionPath)
-				if err != nil {
-					err = fmt.Errorf("error deleting the previous version's folder: %w", err)
-					fmt.Println(err) // Print the error or handle it appropriately
+				if previousVersion != "" {
+					previousVersionPath := filepath.Join(SALTOLocation, previousVersion)
+					err = os.RemoveAll(previousVersionPath)
+					if err != nil {
+						err = fmt.Errorf("error deleting the previous version's folder: %w", err)
+						fmt.Println(err) // Print the error or handle it appropriately
+					}
 				}
 
 				// The previus version is what has been stored in current version
